pyproject.toml
[project]
name = "netops"
version = "0.1.0"
description = "Network operations tooling (Mikrotik, Omada, Speed Audit, etc.)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "paramiko",
  "python-dotenv",
  "click",
  "telnetlib3"
]

[project.scripts]
netops = "netops.cli:cli"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
